Adding Codecov Workflow and Badge (#1)
* Add Codecov workflow in GitHub actions

A new workflow for Codecov has been introduced in the project by adding ".github/workflows/codecov.yml". This workflow is set to run on push and pull request events. It performs several steps, including setting up PHP 8.2, installing dependencies, running tests, collecting coverage, and uploading it to Codecov.

* Add codecov badge to README

The Codecov badge has been added to the README to show test coverage status. This further increases the transparency of the project's testing status and provides a quick overview for all contributors and users.

* add CHANGELOG.md