Update stringmanipulation package version (#2)
* Update stringmanipulation package version

The version of the "marjovanlier/stringmanipulation" package has been updated from "^1.0.54" to "^1.0.59" in the composer.json file. This should bring in the latest updates and bug fixes from the package.

* Update CHANGELOG.md

* Updated codecov workflow to correct the slug

The slug in the codecov workflow file was updated to point to the correct repository 'SouthAfricanIDValidator' instead of the previous 'StringManipulation'. This ensures that coverage reports are sent to the right location.

* Remove duplicate date entry in CHANGELOG.md

This commit removes the duplicate date entry ## 2024-02-18 in the changelog. Maintaining a clean and right changelog is important for overall recording. Every release date should only appear once, with all relevant changes recorded under that date.