(Added) Updated GitHub Actions Workflow and CodeRabbit Configuration (#3)
* Add new .coderabbit.yaml configuration file

This commit introduces a new .coderabbit.yaml configuration file. The file enables early access, configures the code review process and chat settings. It also specifies when to use auto-review, which branch names to consider, and which keywords to ignore in title.

* Updated GitHub Actions Workflow

Expanded PHP version support in the build strategy and introduced the "release" job in the workflow. PHP version support was now extended to both PHP 8.2 and 8.3, with the "release" job added to handle successful build scenarios. These changes enhance test coverage while providing a smoother release process.

* Update CHANGELOG.md

---------

Co-authored-by: codiumai-pr-agent-pro[bot] <151058649+codiumai-pr-agent-pro[bot]@users.noreply.github.com>