(Changed) Update Configuration Settings and Dependency Version (#5)
* Update composer dependencies and streamline script commands

This commit updates the versions of multiple dependencies listed in composer.json for increased functionality and security. Script commands have also been simplified by removing unnecessary parts. This aims to enhance readability and maintainability of test script commands.

* Refine conditional check in GitHub workflow configuration

This commit removes an extra space in the conditional statement that checks the outcome of the 'new-tag' step in our action workflow. This optimization ensures clear code readability and avoids any potential parsing issues.

* Update incremental review settings and changelog push

The commit adjusts the minimal thresholds for incremental reviews, moving from 0 to 2 for commits and from 0 to 10 for minutes. It also changes the setting for pushing changelog changes from true to false to prevent automatic changes.

* Remove CHANGELOG.md file

The `CHANGELOG.md` file has been removed from the project. It seems to be no longer necessary as the important updates and changes are likely managed or documented elsewhere. This helps to reduce clutter and improve the project organization.

* Update stringmanipulation dependency to v1.0.64

The stringmanipulation package used in this project has been upgraded to version 1.0.64 for the latest features and bug fixes. The composer.json file has been edited to reflect this dependency change.

* Remove CHANGELOG.md file

The CHANGELOG.md file was removed from the project. This removal was necessary as it had become outdated, and a more automated, reliable, and up-to-date method of tracking changes (for example, through version control history or release notes) was preferred. The previous method involving manually updating this file was deemed as inefficient and prone to errors.

* Update stringmanipulation package to version 1.0.64

The stringmanipulation package used in the project has been updated from version 1.0.63 to 1.0.64 in the composer.json file. This update will ensure we have the latest features and bug fixes from the library.

* add CHANGELOG.md

* Update CHANGELOG.md

* Remove CHANGELOG.md

The commit removes the CHANGELOG.md file in order to streamline project documentation practices. The removal was previously listed in the changes of the file itself. It is part of the effort to enhance the project's review processes and improve security and functionality.

---------

Co-authored-by: codiumai-pr-agent-pro[bot] <151058649+codiumai-pr-agent-pro[bot]@users.noreply.github.com>